+++ test/libsolidity/smtCheckerTests/operators/shifts/shift_cleanup.sol
@@ -1,0 +1,16 @@
+pragma experimental SMTChecker;
+
+contract C {
+	function f() public pure returns (uint16 x) {
+		x = 0xffff;
+		x += 32;
+		x = x << 8;
+		x = x >> 16;
+		assert(x == 0);
+		// Fails because x = 0.
+		assert(x == 10);
+	}
+}
+// ----
+// Warning 4984: (109-116): Overflow (resulting value larger than 65535) happens here.
+// Warning 6328: (193-208): Assertion violation happens here.